Make getCurrUserId function more globaly accesible
