+++ frontend/src/app/data.service.ts
@@ -6,5 +6,19 @@ import { Observable } from 'rxjs';
   providedIn: 'root'
 })
 export class DataService {
-  
+  getCurrentUserId(): number | null {
+    const userDataString = localStorage.getItem('user');
+    // Retrieve user data from localStorage
+    if (userDataString) {
+      // Parse user data string to an object
+      const userData = JSON.parse(userDataString);
+      
+      // Access the id property
+      return userData.id;
+    } 
+    else {
+      console.error('No user is logged in.');
+      return null;
+    }
+  }
 }
